adding header components with filter and search for översikt
