+++ frontend/src/casedata/components/casedata-filtering/casedata-filtering.component.tsx
@@ -64,76 +64,42 @@ const CaseDataFiltering: React.FC<{
   ownerFilter?: boolean;
   administrators?: (SupportAdmin | Admin)[];
 }> = ({ ownerFilterHandler = () => false, ownerFilter, administrators = [] }) => {
-  const [show, setShow] = useState<boolean>(false);
+  const [show, setShow] = useState<boolean>(true);
   const { selectedErrandStatuses, sidebarLabel, closedErrands } = useAppContext();
   return (
-    <>
-      <div className="flex flex-col w-full gap-24">
-        <div className="w-full flex items-start md:items-center justify-between flex-col md:flex-row gap-12">
-          <h1 className="p-0 m-0">
-            {sidebarLabel || 'Ärenden'}
-            {sidebarLabel === 'Avslutade ärenden' ? ' : ' + closedErrands.totalElements : null}
-          </h1>
-          {/* <div className="w-full md:max-w-[48rem]">
-            <CasedataFilterQuery />
-          </div> */}
-          <Button
-            className="w-full md:w-auto"
-            onClick={() => setShow(!show)}
-            data-cy="Show-filters-button"
-            color="vattjom"
-            variant={show ? 'tertiary' : 'primary'}
-            inverted={show ? false : true}
-            leftIcon={<LucideIcon name="list-filter" size="1.8rem" />}
-          >
-            {show ? 'Dölj filter' : 'Filter'}
-          </Button>
+    <div className="w-fit flex flex-col md:flex-row justify-start items-center p-10 gap-4 flex-wrap">
+      {!isPT() ? (
+        <div className="relative max-md:w-full">
+          <CasedataFilterPropertyDesignation />
         </div>
-
-        <div className={cx(show ? 'visible' : 'hidden')}>
-          <div className="w-full flex flex-col md:flex-row justify-start items-center p-10 gap-4 bg-background-200 rounded-groups flex-wrap">
-            {!isPT() ? (
-              <div className="relative max-md:w-full">
-                <CasedataFilterPropertyDesignation />
-              </div>
-            ) : ['ArendeAvslutat', 'ArendeInkommit', 'Tilldelat'].every((s) => !selectedErrandStatuses.includes(s)) ? (
-              <div className="relative max-md:w-full">
-                <CasedataFilterStatus />
-              </div>
-            ) : null}
-            <div className="relative max-md:w-full">
-              <CasedataFilterCaseType />
-            </div>
-            <div className="relative max-md:w-full">
-              <CasedataFilterPriority />
-            </div>
-            <div className="relative max-md:w-full">
-              <CasedataFilterDates />
-            </div>
-            <div className="relative max-md:w-full">
-              <CasedataFilterAdmins administrators={administrators} />
-            </div>
-            {!isPT() &&
-            ['ArendeAvslutat', 'ArendeInkommit', 'Tilldelat'].every((s) => !selectedErrandStatuses.includes(s)) ? (
-              <div className="relative max-md:w-full">
-                <CasedataFilterStatus />
-              </div>
-            ) : (
-              <div className="relative max-md:w-full">
-                <CasedataFilterPhase />
-              </div>
-            )}
-          </div>
+      ) : ['ArendeAvslutat', 'ArendeInkommit', 'Tilldelat'].every((s) => !selectedErrandStatuses.includes(s)) ? (
+        <div className="relative max-md:w-full">
+          <CasedataFilterStatus />
         </div>
-
-        <div className="flex flex-col md:flex-row justify-start gap-16">
-          <Checkbox data-cy="myErrands-filter" checked={ownerFilter} onChange={() => ownerFilterHandler(!ownerFilter)}>
-            Mina ärenden
-          </Checkbox>
-          <CasedataFilterTags administrators={administrators} />
-        </div>
+      ) : null}
+      <div className="relative max-md:w-full">
+        <CasedataFilterCaseType />
       </div>
-    </>
+      <div className="relative max-md:w-full">
+        <CasedataFilterPriority />
+      </div>
+      <div className="relative max-md:w-full">
+        <CasedataFilterDates />
+      </div>
+      <div className="relative max-md:w-full">
+        <CasedataFilterAdmins administrators={administrators} />
+      </div>
+      {!isPT() &&
+      ['ArendeAvslutat', 'ArendeInkommit', 'Tilldelat'].every((s) => !selectedErrandStatuses.includes(s)) ? (
+        <div className="relative max-md:w-full">
+          <CasedataFilterStatus />
+        </div>
+      ) : (
+        <div className="relative max-md:w-full">
+          <CasedataFilterPhase />
+        </div>
+      )}
+    </div>
   );
 };
