+++ frontend/src/casedata/components/casedata-filtering/components/casedata-filter-query.component.tsx
@@ -34,7 +34,7 @@ export const CasedataFilterQuery: React.FC = () => {
   return (
     <SearchField
       value={query}
-      size={isMobile ? 'md' : 'lg'}
+      size="md"
       data-cy="query-filter"
       onChange={(e) => {
         setQuery(e.target.value);
