+++ frontend/src/common/components/layout/sidebar-layout.component.tsx
@@ -1,18 +1,25 @@
 import { IErrand } from '@casedata/interfaces/errand';
 import { useAppContext } from '@common/contexts/app.context';
 import { User } from '@common/interfaces/user';
-import { getApplicationName } from '@common/services/application-service';
-import { CookieConsent, Link } from '@sk-web-gui/react';
+import { getApplicationEnvironment, getApplicationName, isMEX, isPT } from '@common/services/application-service';
+import { Button, CookieConsent, Header, Link, Logo } from '@sk-web-gui/react';
 import { SupportErrand } from '@supportmanagement/services/support-errand-service';
 import Head from 'next/head';
 import NextLink from 'next/link';
 import { MainErrandsSidebar } from '../main-errands-sidebar/main-errands-sidebar.component';
+import router from 'next/router';
+import { FormProvider, useForm } from 'react-hook-form';
+import { SupportManagementFilterQuery } from '@supportmanagement/components/supportmanagement-filtering/components/supportmanagement-filter-query.component';
+import SupportManagementFiltering, {
+  SupportManagementFilter,
+  SupportManagementValues,
+} from '@supportmanagement/components/supportmanagement-filtering/supportmanagement-filtering.component';
 
 export default function SidebarLayout({ title, children, showAttestationTable, setShowAttestationTable }) {
   const { user, errand, supportErrand }: { user: User; errand: IErrand; supportErrand: SupportErrand } =
     useAppContext();
   const applicationName = getApplicationName();
-
+  const applicationEnvironment = getApplicationEnvironment();
   const hostName = window.location.hostname;
 
   return (
@@ -21,12 +28,14 @@ export default function SidebarLayout({ title, children, showAttestationTable, s
         <title>{title}</title>
         <meta name="description" content={applicationName} />
       </Head>
-      <div className="flex min-h-screen h-full overflow-hidden">
-        <MainErrandsSidebar
-          showAttestationTable={showAttestationTable}
-          setShowAttestationTable={setShowAttestationTable}
-        />{' '}
-        <div className="w-full flex-shirnk flex justify-center pl-[32rem]">{children}</div>
+      <div className="min-h-screen h-full overflow-hidden w-full">
+        <div className="flex h-full overflow-hidden w-full">
+          <MainErrandsSidebar
+            showAttestationTable={showAttestationTable}
+            setShowAttestationTable={setShowAttestationTable}
+          />{' '}
+          <div className="w-full flex-shirnk flex justify-center pl-[32rem]">{children}</div>
+        </div>
       </div>
 
       <CookieConsent
