+++ frontend/src/pages/oversikt.tsx
@@ -9,6 +9,8 @@ import { getSupportMetadata } from '@supportmanagement/services/support-metadata
 import { useRouter } from 'next/router';
 import { useEffect, useRef, useState } from 'react';
 import { AttestationTab } from '@supportmanagement/components/attestation-tab/attestation-tab.component';
+import OngoingCaseDataHeader from '@casedata/components/ongoing-casedata-errands/ongoing-casedata-errands-header.component';
+import OngoingSupportManagementHeader from '@supportmanagement/components/ongoing-support-errands/ongoing-support-errands-header.component';
 
 export const Oversikt: React.FC = () => {
   const router = useRouter();
@@ -44,25 +46,31 @@ export const Oversikt: React.FC = () => {
   return (
     <>
       {isKC() || isIK() || isLOP() ? (
-        <SidebarLayout
-          title={`${getApplicationName()} - Översikt`}
-          setShowAttestationTable={setShowAttestationTable}
-          showAttestationTable={showAttestationTable}
-        >
-          {isLOP() && showAttestationTable && user.permissions.canViewAttestations ? (
-            <AttestationTab />
-          ) : municipalityId ? (
-            <OngoingSupportErrands ongoing={{ errands: [], labels: [] }} />
-          ) : null}
-        </SidebarLayout>
+        <>
+          <OngoingSupportManagementHeader />
+          <SidebarLayout
+            title={`${getApplicationName()} - Översikt`}
+            setShowAttestationTable={setShowAttestationTable}
+            showAttestationTable={showAttestationTable}
+          >
+            {isLOP() && showAttestationTable && user.permissions.canViewAttestations ? (
+              <AttestationTab />
+            ) : municipalityId ? (
+              <OngoingSupportErrands ongoing={{ errands: [], labels: [] }} />
+            ) : null}
+          </SidebarLayout>
+        </>
       ) : (
-        <SidebarLayout
-          setShowAttestationTable={setShowAttestationTable}
-          showAttestationTable={showAttestationTable}
-          title={`${getApplicationName()} - Översikt`}
-        >
-          {isPT() ? <OngoingCaseDataErrands /> : isMEX() ? <OngoingCaseDataErrands /> : <></>}
-        </SidebarLayout>
+        <>
+          <OngoingCaseDataHeader />
+          <SidebarLayout
+            setShowAttestationTable={setShowAttestationTable}
+            showAttestationTable={showAttestationTable}
+            title={`${getApplicationName()} - Översikt`}
+          >
+            {isPT() ? <OngoingCaseDataErrands /> : isMEX() ? <OngoingCaseDataErrands /> : <></>}
+          </SidebarLayout>
+        </>
       )}
     </>
   );
