+++ frontend/src/supportmanagement/components/supportmanagement-filtering/supportmanagement-filtering.component.tsx
@@ -91,88 +91,45 @@ const SupportManagementFiltering: React.FC<{
   ownerFilter?: boolean;
   administrators?: (SupportAdmin | Admin)[];
 }> = ({ ownerFilterHandler = () => false, ownerFilter, administrators = [] }) => {
-  const [show, setShow] = useState<boolean>(false);
-  const {
-    selectedSupportErrandStatuses: selectedSupportErrandStatuses,
-    sidebarLabel,
-    solvedSupportErrands,
-  } = useAppContext();
+  const [show, setShow] = useState<boolean>(true);
 
   return (
-    <>
-      <div className="flex flex-col w-full gap-24">
-        <div className="w-full flex items-start md:items-center justify-between flex-col md:flex-row gap-12">
-          <h1 className="p-0 m-0">
-            {sidebarLabel || 'Ärenden'}
-            {sidebarLabel === 'Avslutade ärenden' ? ' : ' + solvedSupportErrands.totalElements : null}
-          </h1>
-
-          <div className="w-full md:max-w-[48rem]">
-            {/*
-            * TODO needs better API support
-            <SupportManagementFilterQuery />
-            */}
+    <div className="w-fit flex flex-col md:flex-row justify-start items-center p-10 gap-4 flex-wrap">
+      {isLOP() || isIK() ? (
+        <>
+          <div className="relative max-md:w-full">
+            <SupportManagementFilterLabelCategory />
           </div>
-          <Button
-            className="w-full md:w-auto"
-            onClick={() => setShow(!show)}
-            data-cy="show-filters-button"
-            color="vattjom"
-            variant={show ? 'tertiary' : 'primary'}
-            inverted={show ? false : true}
-            leftIcon={<LucideIcon name="list-filter" size="1.8rem" />}
-          >
-            {show ? 'Dölj filter' : 'Filter'}
-          </Button>
-        </div>
-
-        <div className={cx(show ? 'visible' : 'hidden')}>
-          <div className="w-full flex flex-col md:flex-row justify-start items-center p-10 gap-4 bg-background-200 rounded-groups flex-wrap">
-            {isLOP() || isIK() ? (
-              <>
-                <div className="relative max-md:w-full">
-                  <SupportManagementFilterLabelCategory />
-                </div>
-                <div className="relative max-md:w-full">
-                  <SupportManagementFilterLabelType />
-                </div>
-                <div className="relative max-md:w-full">
-                  <SupportManagementFilterLabelSubType />
-                </div>
-              </>
-            ) : (
-              <>
-                <div className="relative max-md:w-full">
-                  <SupportManagementFilterCategory />
-                </div>
-                <div className="relative max-md:w-full">
-                  <SupportManagementFilterType />
-                </div>
-              </>
-            )}
-            <div className="relative max-md:w-full">
-              <SupportManagementFilterPriority />
-            </div>
-            <div className="relative max-md:w-full">
-              <SupportManagementFilterDates />
-            </div>
-            <div className="relative max-md:w-full">
-              <SupportManagementFilterAdmins administrators={administrators} />
-            </div>
-            <div className="relative max-md:w-full">
-              <SupportManagementFilterChannel />
-            </div>
+          <div className="relative max-md:w-full">
+            <SupportManagementFilterLabelType />
           </div>
-        </div>
-
-        <div className="flex flex-col md:flex-row justify-start gap-16">
-          <Checkbox checked={ownerFilter} onChange={() => ownerFilterHandler(!ownerFilter)}>
-            Mina ärenden
-          </Checkbox>
-          <SupportManagementFilterTags administrators={administrators} />
-        </div>
+          <div className="relative max-md:w-full">
+            <SupportManagementFilterLabelSubType />
+          </div>
+        </>
+      ) : (
+        <>
+          <div className="relative max-md:w-full">
+            <SupportManagementFilterCategory />
+          </div>
+          <div className="relative max-md:w-full">
+            <SupportManagementFilterType />
+          </div>
+        </>
+      )}
+      <div className="relative max-md:w-full">
+        <SupportManagementFilterPriority />
       </div>
-    </>
+      <div className="relative max-md:w-full">
+        <SupportManagementFilterDates />
+      </div>
+      <div className="relative max-md:w-full">
+        <SupportManagementFilterAdmins administrators={administrators} />
+      </div>
+      <div className="relative max-md:w-full">
+        <SupportManagementFilterChannel />
+      </div>
+    </div>
   );
 };
